move sentry
