+++ packages/core/src/modules/goody.ts
@@ -1,4 +1,4 @@
-import { reportException } from '@/modules/sentry/use-cases/report-exception';
+import { reportException } from '@/infrastructure/sentry';
 import { fail, type Result, success } from '@/shared/utils/core.utils';
 
 // Constants
