+++ packages/core/src/modules/member/events/member-primary-email-changed.ts
@@ -2,11 +2,11 @@ import { db } from '@oyster/db';
 
 import { job } from '@/infrastructure/bull';
 import { type GetBullJobData } from '@/infrastructure/bull.types';
+import { reportException } from '@/infrastructure/sentry';
 import {
   AIRTABLE_FAMILY_BASE_ID,
   AIRTABLE_MEMBERS_TABLE_ID,
 } from '@/modules/airtable/airtable.core';
-import { reportException } from '@/modules/sentry/use-cases/report-exception';
 import {
   getSlackUserByEmail,
   updateSlackEmail,
