+++ packages/core/src/modules/shopify.ts
@@ -1,4 +1,4 @@
-import { reportException } from '@/modules/sentry/use-cases/report-exception';
+import { reportException } from '@/infrastructure/sentry';
 import { fail, type Result, success } from '@/shared/utils/core.utils';
 import { RateLimiter } from '@/shared/utils/rate-limiter';
