+++ packages/core/src/modules/slack/services/slack-admin.service.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 
 import { job } from '@/infrastructure/bull';
 import { redis } from '@/infrastructure/redis';
-import { reportException } from '@/modules/sentry/use-cases/report-exception';
+import { reportException } from '@/infrastructure/sentry';
 import { ErrorWithContext, ZodParseError } from '@/shared/errors';
 import { RateLimiter } from '@/shared/utils/rate-limiter';
